error handling add-on
error handling was added to the indexHandler function to ensure that the web server works correctly even in unexpected situations. If any error occurs when loading the HTML template, an HTTP 500 error (Internal Server Error) is returned and an error message is displayed to the user.

In addition, the log.Fatal function is used to handle any errors that occur when starting the web server. This ensures that the program will stop in case of a critical error instead of continuing to run with undefined behaviour.